+++ src/Tray/Program.cs
@@ -112,7 +112,7 @@ internal static class Program
         {
             string path = args.Length >= 2 ? args[1] : "docs/social-preview.png";
             using Bitmap bmp = IconRenderer.RenderSocial(1280, 640);
-            bmp.Save(path, System.Drawing.Imaging.ImageFormat.Png);
+            using (FileStream fs = OutFile.Create(path)) bmp.Save(fs, System.Drawing.Imaging.ImageFormat.Png);
             Console.WriteLine("wrote " + Path.GetFullPath(path));
             return;
         }