+++ src/Ui/SettingsPage.xaml.cs
@@ -328,7 +328,7 @@ internal partial class SettingsPage : System.Windows.Controls.UserControl
 
         var encoder = new System.Windows.Media.Imaging.PngBitmapEncoder();
         encoder.Frames.Add(System.Windows.Media.Imaging.BitmapFrame.Create(rtb));
-        using System.IO.FileStream fs = System.IO.File.Create(path);
+        using FileStream fs = OutFile.Create(path);
         encoder.Save(fs);
     }